refactor(components): оптимизация кода за счет хука useMovie
