+++ src/components/app/app.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import MainPage from "../pages/main/main";
 import NotFound from "../pages/not-found/not-found";
 import MyList from "../pages/my-list/my-list";
-import PlayerPage from "../pages/player-page/player-page";
+import Player from "../pages/player/player";
 import SignIn from "../pages/sign-in/sign-in";
 import MoviePage from "../pages/movie-page/movie-page";
 import AddReview from "../pages/add-review/add-review";
@@ -18,18 +18,9 @@ const App = ({movieItems, genresItems, promoMovie}) => {
         }/>
         <Route path={RoutePath.LOGIN} exact component={SignIn}/>
         <Route path={RoutePath.MY_LIST} exact component={MyList}/>
-        <Route path={RoutePath.PLAYER_ID} component={PlayerPage}/>
-        <Route path={RoutePath.FILM_ID} exact
-          render={
-            ({match}) => {
-              const {id} = match.params;
-              return <MoviePage itemId={Number(id)} />;
-            }}/>
-        <Route path={RoutePath.FILM_REVIEW} render={
-          ({match}) => {
-            const {id} = match.params;
-            return <AddReview itemId={Number(id)} />;
-          }}/>
+        <Route path={RoutePath.PLAYER_ID} component={Player}/>
+        <Route path={RoutePath.FILM_ID} exact component={MoviePage}/>
+        <Route path={RoutePath.FILM_REVIEW} component={AddReview}/>
         <Route component={NotFound}/>
       </Switch>
     </Router>
