+++ src/components/hooks/use-movie/useMovie.jsx
@@ -1,7 +1,11 @@
 import {useState, useEffect} from "react";
 import {movieItems} from "../../../mocks/movie-items";
+import {useParams} from "react-router-dom";
 
-const useMovie = (itemId) => {
+const useMovie = () => {
+  const {id} = useParams();
+  const itemId = Number(id);
+
   const movie = movieItems.find((item) => item.id === itemId);
 
   const [currentMovie, setCurrentMovie] = useState(movie);
